finally

+++ crud/app/Http/Controllers/CourseController.php
@@ -38,7 +38,14 @@ class CourseController extends Controller
 
     public function create() 
     {
-        return view('admin/courses/new');
+        $user = Auth::User();
+
+        if($user->admin){
+            return view('admin/courses/new');
+        }else{
+            \Session::flash('status', 'Você não tem permissão para acessar essa página!');
+            return redirect('course');
+        }
     }
 
     public function store(Request $request) 
@@ -59,32 +66,54 @@ class CourseController extends Controller
 
     public function destroy($id)
      {
-        $p = Course::findOrFail($id);
-        $p->delete();
+        $user = Auth::User();
 
-        \Session::flash('status', 'Curso excluído com sucesso.');
-        return redirect('course');
+        if($user->admin){
+            $p = Course::findOrFail($id);
+            $p->delete();
+
+            \Session::flash('status', 'Curso excluído com sucesso.');
+            return redirect('course');
+        }else{
+            \Session::flash('status', 'Você não tem permissão para acessar essa página!.');
+            return redirect('course');
+        }
+        
     }
 
     public function edit($id) {
-        $course = Course::findOrFail($id);
-        
 
-        return view('admin/courses.edit', ['course' => $course]);
+        $user = Auth::User();
+
+        if($user->admin){
+            $course = Course::findOrFail($id);
+            
+            return view('admin/courses.edit', ['course' => $course]);
+        }else{
+            \Session::flash('status', 'Você não tem permissão para acessar essa página!.');
+            return redirect('course');
+        }
     }
 
     public function update(Request $request, $id) {
-        $p = Course::findOrFail($id);
-        $p->nome = $request->input('nome');
-        $p->ementa = $request->input('ementa');
-        $p->qtdAlunos = $request->input('qtdAlunos');
-        
-        if ($p->save()) {
-            \Session::flash('status', 'Curso atualizado com sucesso!');
-            return redirect('course');
-        } else {
-            \Session::flash('status', 'Ocorreu um erro ao atualizar o Curso.');
-            return view('courses.edit');
+        $user = Auth::User();
+
+        if($user->admin){
+            $p = Course::findOrFail($id);
+            $p->nome = $request->input('nome');
+            $p->ementa = $request->input('ementa');
+            $p->qtdAlunos = $request->input('qtdAlunos');
+            
+            if ($p->save()) {
+                \Session::flash('status', 'Curso atualizado com sucesso!');
+                return redirect('course');
+            } else {
+                \Session::flash('status', 'Ocorreu um erro ao atualizar o Curso.');
+                return view('courses.edit');
+            }
+        }else{
+            \Session::flash('status', 'Você não tem permissão para acessar essa página!.');
+            return redirect('course'); 
         }
     }
