+++ crud/app/Http/Controllers/StudentController.php
@@ -34,35 +34,56 @@ class StudentController extends Controller
 
     public function create() 
     {
-        $Student = User::all();
-        return view('student/new', ['student' => $Student]);
+        $user = Auth::User();
+
+        if($user->admin){
+            $Student = User::all();
+            return view('student/new', ['student' => $Student]);
+        }else{
+            \Session::flash('status', 'Você não tem permissão para acessar essa página!.');
+            return redirect('course');
+        }
     }
 
     public function store(Request $request) 
     {
-        $p = new student;
-        $p->nome = $request->input('nome');
-        $p->cpf = $request->input('cpf');
-        $p->rg = $request->input('rg');
-        $p->endereco = $request->input('address');
-        $p->rg = $request->input('cellphone');
-        
-        if ($p->save()) {
-            \Session::flash('status', 'Estudante criado com sucesso!');
-            return redirect('/student');
-        } else {
-            \Session::flash('status', 'Ocorreu um erro ao criar o estudante!');
-            return view('student.new');
+        $user = Auth::User();
+
+        if($user->admin){
+            $p = new student;
+            $p->nome = $request->input('nome');
+            $p->cpf = $request->input('cpf');
+            $p->rg = $request->input('rg');
+            $p->endereco = $request->input('address');
+            $p->rg = $request->input('cellphone');
+            
+            if ($p->save()) {
+                \Session::flash('status', 'Estudante criado com sucesso!');
+                return redirect('/student');
+            } else {
+                \Session::flash('status', 'Ocorreu um erro ao criar o estudante!');
+                return view('student.new');
+            }
+        }else{
+            \Session::flash('status', 'Você não tem permissão para acessar essa página!.');
+            return redirect('course');
         }
     }
 
     public function destroy($id)
      {
-        $p = User::findOrFail($id);
-        $p->delete();
+        $user = Auth::User();
 
-        \Session::flash('status', 'Estudante excluído com sucesso.');
-        return redirect('student');
+        if($user->admin){
+            $p = User::findOrFail($id);
+            $p->delete();
+
+            \Session::flash('status', 'Estudante excluído com sucesso.');
+            return redirect('student');
+        }else{
+            \Session::flash('status', 'Você não tem permissão para acessar essa página!.');
+            return redirect('course');
+        }
     }
 
     public function update(Request $request, $id) {
@@ -80,6 +101,22 @@ class StudentController extends Controller
         } else {
             \Session::flash('status', 'Ocorreu um erro ao atualizar seus dados.');
             return view('student');
+        }
+    }
+
+    public function updateUser($id)
+    {
+        $user = User::findOrFail($id);    
+        if ($user->admin==true){
+            $user->admin = false;
+            $user->save();
+            \Session::flash('status', 'Usuário se tornou um aluno!');
+            return redirect('/student');
+        }else{
+            $user->admin = true;
+            $user->save();
+            \Session::flash('status', 'O usuário tornou-se um administrador');
+            return redirect('/student');
         }
     }
